Clean up estimators
